Added support for orders with non default currency
If the Currency is allowed for the store it will be set on the order, if the store doesn't allow the currency the default is used instead. This stops needing a storeview for every currency you sell products in. After the order is placed the currency is set back to the default to stop any potential issues because of caching.