Ensure shipping cost is always a float
When calculating the shipping costs, the given value from the JSON
request can also be a string, this will cause an error since the return
type is set to be a float. To ensure this will not give any issues, the
shipping costs variable is casted to a float in all cases.