Change some methods to public to be accessible for plugins
There are a lot of private functions and in my opinion some of them should be public to make them accessible for plugins in Magento. In this PR case I changed just two methods because I need them to be accessible to add some changes on client level and I think they don't need to be private.